Break up info page into component
